requirements.txt
-e git+git@github.com:ctsit/cappy.git@7824ef278dd47b61efe01320409988230807642a#egg=cappy
certifi==2017.7.27.1
chardet==3.0.4
docopt==0.6.2
idna==2.6
python-cappy==0.0.2
PyYAML==3.12
requests==2.18.4
urllib3==1.22
